Refactored user endpoint into controllers
